Move the modal grid's inline stars and trend glyphs into modules/icons as SparklesMorpho and TrendingUp (PR #1773 review), matching the convention LightningFilled set in this PR; SparklesMorpho keeps its useId gradient defs (the widgets Sparkles bakes static ids and a different fill), and GradientRing stays local — it is a ring wrapper around children, not an icon.
