+++ apps/webapp/src/modules/icons/index.ts
@@ -48,6 +48,8 @@ import { Wallet } from './Wallet';
 import { Pendle } from './Pendle';
 import { Merkl } from './Merkl';
 import { StarsFilled } from './StarsFilled';
+import { SparklesMorpho } from './SparklesMorpho';
+import { TrendingUp } from './TrendingUp';
 import { IllustrationStaked } from './IllustrationStaked';
 import { IllustrationStakingLogomark } from './IllustrationStakingLogomark';
 
@@ -102,6 +104,8 @@ export {
   Pendle,
   Merkl,
   StarsFilled,
+  SparklesMorpho,
+  TrendingUp,
   IllustrationStaked,
   IllustrationStakingLogomark
 };
